+++ .eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  rules: {
+    eqeqeq: 'off',
+    'no-redeclare': 'off',
+    '@typescript-eslint/no-redeclare': ['error', { ignoreDeclarationMerge: true }],
+  },
+};