+++ src/helpers.ts
@@ -1,56 +1,137 @@
-import { ARRAY, BOOLEAN, NUMBER, OBJECT, STRING } from './constants';
-import { UnknownKey, safeSpread } from './utils';
 import {
-  ArrayValidator,
+  O,
+  R,
   BooleanValidator,
+  ListValidator,
+  ListofValidator,
   NumberValidator,
-  ObjectValidator,
+  RecordValidator,
+  RecordofValidator,
+  Schema,
   StringValidator,
   Validator,
+  BooleanOptions,
+  ListOptions,
+  ListofOptions,
+  NumberOptions,
+  RecordOptions,
+  RecordofOptions,
+  StringOptions,
 } from './validatorTypes';
+import { $boolean, $list, $listof, $number, $record, $recordof, $string } from './constants';
 
-type WithOutType<T> = Omit<T, 'type'>;
+export function string(): R<StringOptions>;
+export function string(config: Omit<StringOptions, 'type'>): R<StringOptions>;
+export function string(config: { optional: false } & Omit<StringOptions, 'type'>): R<StringOptions>;
+export function string(config: { optional: true } & Omit<StringOptions, 'type'>): O<StringOptions>;
 
-interface ValidatorsHelpers {
-  string: (opts?: Partial<WithOutType<StringValidator>>) => StringValidator;
-  bool: (opts?: Partial<WithOutType<BooleanValidator>>) => BooleanValidator;
-  number: (opts?: Partial<WithOutType<NumberValidator>>) => NumberValidator;
-  record: (v: Record<string, Validator>, opts?: { optional: boolean }) => ObjectValidator;
-  recordof: (v: Validator, opts?: { optional: boolean }) => ObjectValidator;
-  listof: (v: Validator, opts?: { optional: boolean }) => ArrayValidator;
+export function string(
+  config?: { optional?: boolean } & Omit<StringOptions, 'type'>
+): StringValidator {
+  return {
+    type: $string,
+    optional: !!config?.optional,
+    ...config,
+  };
 }
 
-const optional = (value: boolean | void) => (typeof value == 'boolean' ? value : false);
+export function number(): R<NumberOptions>;
+export function number(config: Omit<NumberOptions, 'type'>): R<NumberOptions>;
+export function number(config: { optional: true } & Omit<NumberOptions, 'type'>): O<NumberOptions>;
+export function number(config: { optional: false } & Omit<NumberOptions, 'type'>): R<NumberOptions>;
 
-export const _: ValidatorsHelpers = {
-  string(opts) {
-    return { ...safeSpread(opts), type: STRING, optional: optional(opts?.optional) };
-  },
-  bool(opts) {
-    return { ...safeSpread(opts), type: BOOLEAN, optional: optional(opts?.optional) };
-  },
-  number(opts) {
-    return { ...safeSpread(opts), type: NUMBER, optional: optional(opts?.optional) };
-  },
-  listof(v, opts) {
-    return { ...safeSpread(opts), type: ARRAY, of: v, optional: optional(opts?.optional) };
-  },
-  recordof(v, opts) {
-    return {
-      ...safeSpread(opts),
-      optional: optional(opts?.optional),
-      type: OBJECT,
-      shape: {
-        [UnknownKey]: v,
-      },
-    };
-  },
-  record(v, opts) {
-    return {
-      ...safeSpread(opts),
-      optional: optional(opts?.optional),
-      type: OBJECT,
-      shape: v,
-    };
-  },
-};
+export function number(
+  config?: { optional?: boolean } & Omit<NumberOptions, 'type'>
+): NumberValidator {
+  return {
+    type: $number,
+    optional: !!config?.optional,
+    ...config,
+  };
+}
+
+export function boolean(): R<BooleanOptions>;
+export function boolean(config: { optional: true }): O<BooleanOptions>;
+export function boolean(config: { optional: false }): R<BooleanOptions>;
+
+export function boolean(config?: { optional: boolean }): BooleanValidator {
+  return {
+    type: $boolean,
+    optional: !!config?.optional,
+  };
+}
+
+export function list<T extends R<Validator>[]>(list: T): R<ListOptions<T>>;
+export function list<T extends R<Validator>[]>(
+  list: T,
+  config: { optional: false }
+): R<ListOptions<T>>;
+export function list<T extends R<Validator>[]>(
+  list: T,
+  config: { optional: true }
+): O<ListOptions<T>>;
+
+export function list<T extends R<Validator>[]>(
+  list: T,
+  config?: { optional: boolean }
+): ListValidator<T> {
+  return {
+    type: $list,
+    optional: !!config?.optional,
+    shape: list.map(v => ({ ...v, optional: false })) as T,
+  };
+}
+
+export function listof<T extends R<Validator>>(v: T): R<ListofOptions<T>>;
+export function listof<T extends R<Validator>>(
+  v: T,
+  config: { optional: false }
+): R<ListofOptions<T>>;
+export function listof<T extends R<Validator>>(
+  v: T,
+  config: { optional: true }
+): O<ListofOptions<T>>;
+
+export function listof<T extends R<Validator>>(
+  v: T,
+  config?: { optional: boolean }
+): ListofValidator<T> {
+  return {
+    type: $listof,
+    optional: !!config?.optional,
+    of: { ...v, optional: false },
+  };
+}
+
+export function record<T extends Schema>(s: T): R<RecordOptions<T>>;
+export function record<T extends Schema>(s: T, config: { optional: false }): R<RecordOptions<T>>;
+export function record<T extends Schema>(s: T, config: { optional: true }): O<RecordOptions<T>>;
+
+export function record<T extends Schema>(s: T, config?: { optional: boolean }): RecordValidator<T> {
+  return {
+    type: $record,
+    optional: !!config?.optional,
+    shape: s,
+  };
+}
+
+export function recordof<T extends R<Validator>>(v: T): R<RecordofOptions<T>>;
+export function recordof<T extends R<Validator>>(
+  v: T,
+  config: { optional: false }
+): R<RecordofOptions<T>>;
+export function recordof<T extends R<Validator>>(
+  v: T,
+  config: { optional: true }
+): O<RecordofOptions<T>>;
+
+export function recordof<T extends R<Validator>>(
+  v: T,
+  config?: { optional: boolean }
+): RecordofValidator<T> {
+  return {
+    type: $recordof,
+    of: { ...v, optional: false },
+    optional: !!config?.optional,
+  };
+}