+++ src/index.ts
@@ -1,2 +1,5 @@
+import * as helpers from './helpers';
+
+export * from './validatorTypes';
 export * from './createSchema';
-export * from './helpers';
+export const _ = helpers;